v.1.94

+++ Firmware/Source code/Hardware Version 3.x (HW3.2)/NixieClockShield_NCS314/doIndication314_HW3.x.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 unsigned int SymbolArray[10]={1, 2, 4, 8, 16, 32, 64, 128, 256, 512};
 //unsigned int SymbolArray[10]={0xFFFE, 0xFFFD, 0xFFFB, 0xFFF7, 0xFFEF, 0xFFDF, 0xFFBF, 0xFF7F, 0xFEFF, 0xFDFF};
-const unsigned int fpsLimit=16666*3;
+const uint16_t fpsLimit=16666*3;
 const byte LEpin=10;
 
 #endif
